Move Beanstalk driver to new namespace
